+++ rebootpy/__init__.py
@@ -37,7 +37,8 @@ from .message import FriendMessage, PartyMessage
 from .party import (DefaultPartyConfig, DefaultPartyMemberConfig, PartyMember,
                     ClientPartyMember, Party,  ClientParty,
                     ReceivedPartyInvitation, SentPartyInvitation,
-                    PartyJoinConfirmation, PartyJoinRequest, SquadAssignment)
+                    PartyJoinConfirmation, PartyJoinRequest, SquadAssignment,
+                    PlaylistRequest)
 from .presence import Presence, PresenceGameplayStats, PresenceParty
 from .user import (ClientUser, User, BlockedUser, ExternalAuth,
                    UserSearchEntry, SacSearchEntryUser)